npm installed promise base http client, axios
implemented a get emailPreferences and save emailPreferences function that used the api created in the background to get and post to a mongo database.

Added new field to model called userId. This will be used to save and retrieve the user's existing and new email preference. userId will probably not be needed in the future when all of the user portal back end is completed.